refactor(block): refactor `block` package
