+++ src/evm/mod.rs
@@ -1,3 +1,6 @@
+pub mod alloy;
 pub mod api;
+pub mod config;
 pub mod evm;
+pub mod factory;
 pub mod handler;
