--- src/factory/mod.rs
@@ -1,7 +1,0 @@
-pub mod alloy;
-pub mod assembler;
-pub mod block;
-pub mod builder;
-pub mod config;
-pub mod executor;
-pub mod factory;
